% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction_dimension.R
\name{reduction_dimension}
\alias{reduction_dimension}
\title{Calculate the urban vulnerability index for census data}
\usage{
reduction_dimension(ipc)
}
\arguments{
\item{CC}{A matrix or data.frame whose rows sum to one or one hundred.}
}
\value{
A matrix composed of the log ratios values.
}
\description{
Calculate the urban vulnerability index for census data
}
\examples{
C = matrix(1:20,4)/rowSums(matrix(1:20,4))
logratio(C)
}
