% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VulIndex.R
\name{VulIndex}
\alias{VulIndex}
\title{Calculate the urban vulnerability index for census data}
\usage{
VulIndex(id.mun, bairros, entorno, dom.i, dom.ii, pessoa, dom.renda)
}
\arguments{
\item{CC}{A matrix or data.frame whose rows sum to one or one hundred.}
}
\value{
A matrix composed of the log ratios values.
}
\description{
Calculate the urban vulnerability index for census data
}
\examples{
C = matrix(1:20,4)/rowSums(matrix(1:20,4))
logratio(C)
}
